// src/app/models/quiz-model.ts
export class QuizModel {
    question1: string = '';
    question2: string = '';
    question3: string = '';
    question4: string = '';
    question5: string = '';
    question6: string = '';
    question7: string = '';
    question8: string = '';
    question9: string = '';


    constructor() {
         this.question1=this.question1;
         this.question2=this.question2;
         this.question3=this.question3;
         this.question4=this.question4;
         this.question5=this.question5;
         this.question6=this.question6;
         this.question7=this.question7;
         this.question8=this.question8;
         this.question9=this.question9;
    }

}